+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 Livedata SDK is a client library that enables easier integration with the Livescout XML feed. SDK exposes XML feed service interface in a more user-friendly way and isolates the client from having to do XML feed parsing, proper connection handling, error recovery, event queuing and dispatching. 
 It also makes a client solution more stable and robust when it comes to feed handling, especially with the release of new and updated XML feed version.* ScoutFeedType extended with new type - PARTIAL
 
+**2.0.12 (2025-3-**)**
+* finalconfidence property for ScoutEventEntity is changed from string to integer
+* PlayerStatisticsEntity extended with new properties:
+  - homePlayerStatsSet1
+  - homePlayerStatsSet2
+  - homePlayerStatsSet3
+  - homePlayerStatsSet4
+  - homePlayerStatsSet5
+  - awayPlayerStatsSet1
+  - awayPlayerStatsSet2
+  - awayPlayerStatsSet3
+  - awayPlayerStatsSet4
+  - awayPlayerStatsSet5
+* StaticsEntity extended with new TeamPlayerStatsEntity
+
 **2.0.11 (2024-10-18)**
 * MatchUpdateEntity fields became nullable:
   - subteams